requirements.txt
# TODO: Add the versions
flask-cors
Flask-Session
gunicorn
requests
cryptography
setuptools
typing-extensions